+++ src/mailman/docs/NEWS.rst
@@ -24,6 +24,8 @@ Architecture
 Interfaces
 ----------
  * Add property `IUserManager.members` to return all `IMembers` in the system.
+ * Add property `IListmanager.name_compoments` which returns 2-tuples for
+   every mailing list as (list_name, mail_host).
 
 Commands
 --------